fix: corrects PS3 triggers; feature: implements PS3 L1/L2 pressure sensitive buttons for simulated GC buttons; fix: corrects timing for led and rumble setting across Sony DS3/DS4/DS5 and GameCube adapter

+++ PCEMouse/src/devices/device_interface.h
@@ -8,7 +8,7 @@ typedef struct {
     const char* name; // Add this line
     bool (*is_device)(uint16_t vid, uint16_t pid);
     void (*process)(uint8_t dev_addr, uint8_t instance, const uint8_t *report, uint16_t len);
-    void (*task)(uint8_t dev_addr, uint8_t instance, uint8_t player_index, uint8_t rumble);
+    void (*task)(uint8_t dev_addr, uint8_t instance, int player_index, uint8_t rumble);
     bool (*init)(uint8_t dev_addr, uint8_t instance);
     void (*unmount)(uint8_t dev_addr, uint8_t instance);
     // Add other common functions as needed
